backend-21/pretups/src/main/java/com/restapi/superadmin/requestVO/TransferProfileLoadReqVO.java
package com.restapi.superadmin.requestVO;

public class TransferProfileLoadReqVO {
	
	private String domainCode;
	private String categoryCode;
	private String status;
	private String networkCode;
	
	public String getDomainCode() {
		return domainCode;
	}

	public void setDomainCode(String domainCode) {
		this.domainCode = domainCode;
	}

	public String getCategoryCode() {
		return categoryCode;
	}

	public void setCategoryCode(String categoryCode) {
		this.categoryCode = categoryCode;
	}

	public String getStatus() {
		return status;
	}

	public void setStatus(String status) {
		this.status = status;
	}

	public TransferProfileLoadReqVO() {
		
	}

	public String getNetworkCode() {
		return networkCode;
	}

	public void setNetworkCode(String networkCode) {
		this.networkCode = networkCode;
	}
	

}
